[ refactor ] : 모달 에서 리셋 기능
